requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
black==23.1.0
celery==5.2.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.1.5
django-cache-url==3.4.4
django-countries==7.5.1
django-heroku==0.3.1
django-redis==5.2.0
environs==8.0.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
marshmallow==3.19.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
airtable==0.4.8
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.1.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
wrapt==1.15.0
django_heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf_yasg==1.21.5
django-cors-headers==3.14.0
cloudinary==1.32.0
django-cloudinary-storage==0.3.0
django-filter==23.1
django-simple-mail==2.3.5
python-dateutil==2.8.2
currencyapicom
djangorestframework-bulk==0.2.1
rsa==4.9
django-celery-results
sentry-sdk
fcm-django==1.0.14
boto3==1.26.59
django-storages==1.13.2
django-timezone-field==5.0
tzdata==2023.3
django-allauth==0.54.0
dj-rest-auth==4.0.1
app-store-notifications-v2-validator==0.0.8
inapppy==2.5.2
